solune/backend/pyproject.toml
[project]
name = "solune-backend"
version = "0.1.0"
description = "FastAPI backend for Solune"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0,<1",
    "uvicorn[standard]>=0.44.0,<1",
    "githubkit>=0.14.6,<1",
    "httpx>=0.28.0,<1",
    "pydantic>=2.12.0,<3",
    "pydantic-settings>=2.13.0,<3",
    "python-multipart>=0.0.26,<1",
    "pyyaml>=6.0.3,<7",
    # GitHub Copilot SDK (current: v0.x; v2 upgrade to copilot-sdk>=1.0.17 planned)
    "github-copilot-sdk>=0.1.30,<1",
    # Azure OpenAI (optional fallback provider)
    "openai>=2.26.0,<3",
    "azure-ai-inference>=1.0.0b9,<2",
    # Microsoft Agent Framework (v0.2.0 — intelligent chat agent)
    "agent-framework-core>=1.0.1",
    "agent-framework-azure-ai>=1.0.0b1",
    "agent-framework-github-copilot>=1.0.0b1",
    # SSE streaming for real-time agent responses
    "sse-starlette>=3.0.0,<4",
    # SQLite async (durable settings/session storage)
    "aiosqlite>=0.22.0,<1",
    # Encryption for session tokens
    "cryptography>=46.0.5,<47",
    # Signal messaging integration
    "tenacity>=9.1.0,<10",
    "websockets>=16.0,<17",
    # Rate limiting middleware (FastAPI-compatible)
    "slowapi>=0.1.9,<1",
    # GitHub Secrets API encryption (libsodium sealed box)
    "pynacl>=1.6.2,<2",
    # OpenTelemetry — opt-in distributed tracing and metrics (Phase 5)
    "opentelemetry-api>=1.33.0,<2",
    "opentelemetry-sdk>=1.33.0,<2",
    "opentelemetry-instrumentation-fastapi>=0.54b0,<1",
    "opentelemetry-instrumentation-httpx>=0.54b0,<1",
    "opentelemetry-instrumentation-sqlite3>=0.54b0,<1",
    "opentelemetry-exporter-otlp>=1.33.0,<2",
    # Sentry — opt-in error tracking (Phase 5)
    "sentry-sdk[fastapi]>=2.22.0,<3",
    # MCP (Model Context Protocol) server — opt-in (Phase 6 / v0.4.0)
    "mcp>=1.26.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-randomly>=4.0.1",
    "pytest-repeat>=0.9.4",
    "pytest-timeout>=2.2",
    "ruff>=0.15.0",
    "pyright>=1.1.408",
    "hypothesis>=6.131.0",
    "mutmut>=3.5.0",
    "pip-audit>=2.10.0",
    "bandit>=1.9.4",
    "freezegun>=1.5.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Allow prereleases for azure-ai-inference (1.0.0b9) and
# opentelemetry-instrumentation-* (0.54b0) preview packages.
prerelease = "allow"
required-version = ">=0.10.9"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = ["htmlcov"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "BLE",  # flake8-blind-except (ADR-008)
    "C4",   # flake8-comprehensions
    "S110", # try-except-pass detector (non-waivable)
    "TRY",  # tryceratops (exception anti-patterns)
    "UP",   # pyupgrade
    "FURB", # refurb (more Pythonic alternatives)
    "PTH",  # pathlib (prefer pathlib over os.path)
    "PERF", # Perflint (performance anti-patterns)
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # reason: line length enforced by ruff format, not lint
    "TRY400", # reason: higher noise — logging.exception in except blocks
    "TRY401", # reason: higher noise — verbose log messages in except blocks
]

[tool.ruff.lint.per-file-ignores]
# Tests may use broad exceptions and long error messages freely (ADR-008 §Tests).
"tests/**/*.py" = ["BLE001", "TRY003", "TRY301"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require external services)",
    "performance: marks tests as performance tests (require running backend + credentials)",
]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
include = ["src"]
exclude = ["**/__pycache__", "htmlcov"]
reportMissingTypeStubs = false
reportMissingImports = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportMissingParameterType = "error"
reportUnknownParameterType = "warning"
reportUnknownMemberType = "warning"
strict = ["src/api", "src/models", "src/services/agents"]
stubPath = "src/typestubs"

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
fail_under = 75
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[tool.bandit]
targets = ["src"]
severity = "medium"
confidence = "medium"
exclude_dirs = ["tests"]
skips = ["B104"]

[tool.mutmut]
paths_to_mutate = [
    "src/api/",
    "src/middleware/",
    "src/utils.py",
]
tests_dir = ["tests/"]
debug = true
also_copy = [
    "scripts/",
    "src/api/",
    "src/attachment_formatter.py",
    "src/config.py",
    "src/constants.py",
    "src/dependencies.py",
    "src/exceptions.py",
    "src/logging_utils.py",
    "src/main.py",
    "src/middleware/",
    "src/migrations/",
    "src/models/",
    "src/prompts/",
    "src/protocols.py",
    "src/services/",
    "src/utils.py",
    "templates/",
]
